Fix TestMapReference needed an ordered mapping
In 405d5d63b1cfd47b220f95669377e2d51b460cc2 we changed for an ordered
list to a map, however the test is order dependent go maps are
randomized.

Change the test to use a slice with an internal type (instead of pulling
back in k8s.io/kubernetes).

Without this change this test will fail occasionally and has no
guarentee for success because of the random order of maps.